chatbot_files/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
faiss-cpu
sentence-transformers
numpy
pydantic
